Seperate functions into their own file
